users table | delete

+++ app/Http/Controllers/UserController.php
@@ -51,6 +51,10 @@ class UserController extends Controller
         return $user;
     }
 
+    public function getUsers(){
+        return User::all();
+    }
+
     public function editProfile(Request $request){
         $id = JWTAuth::user()->id;
 
@@ -87,12 +91,21 @@ class UserController extends Controller
     }
 
     public function deleteUser($id){
+        $adminId = JWTAuth::user()->id;
         $user = User::find($id);
-        $user->forceDelete();
-        return response()->json([
-            'status' => 'success',
-            'message' => 'Zmazanie pouzivatela bolo uspesne.'
-        ],200);
+        if($id == $adminId){
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Nie je možné odstrániť sám seba.'
+            ],401);
+        }
+        else{
+            $user->forceDelete();
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Zmazanie pouzivatela bolo uspesne.'
+            ],200);
+        }
     }
 
     public function changePassword(Request $request){
